smaller CPU binary
